+++ Sources/StreamChatUI/ChatMessageList/ChatMessageListVC.swift
@@ -327,7 +327,7 @@ open class ChatMessageListVC: _ViewController,
     ///
     /// By default, this will trigger the swipe to reply gesture recognition.
     @objc open func handlePan(_ gesture: UIPanGestureRecognizer) {
-        let canReply = dataSource?.channel(for: self)?.ownCapabilities.contains(.sendReply) ?? false
+        let canReply = dataSource?.channel(for: self)?.canSendReply ?? false
         let isSwipeToReplyEnabled = components.messageSwipeToReplyEnabled
         if canReply && isSwipeToReplyEnabled {
             swipeToReplyGestureHandler.handle(gesture: gesture)
@@ -368,12 +368,12 @@ open class ChatMessageListVC: _ViewController,
 
         let actionsController = components.messageActionsVC.init()
         actionsController.messageController = messageController
-        actionsController.channelConfig = dataSource?.channel(for: self)?.config
+        actionsController.channel = dataSource?.channel(for: self)
         actionsController.delegate = self
 
         let reactionsController: ChatMessageReactionsPickerVC? = {
             guard message.localState == nil else { return nil }
-            guard dataSource?.channel(for: self)?.config.reactionsEnabled == true else {
+            guard dataSource?.channel(for: self)?.canSendReaction == true else {
                 return nil
             }
 
@@ -463,7 +463,7 @@ open class ChatMessageListVC: _ViewController,
 
         messageActionsVC = components.messageActionsVC.init()
         messageActionsVC?.messageController = messageController
-        messageActionsVC?.channelConfig = dataSource?.channel(for: self)?.config
+        messageActionsVC?.channel = dataSource?.channel(for: self)
         messageActionsVC?.delegate = self
 
         guard let messageActions = messageActionsVC?.messageActionsForAlertMenu else {